requirements.txt
wxPython
google-generativeai
requests